+++ backend/src/migrations/004_create_investigation_logs_table.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely';
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('investigation_logs')
-    .addColumn('id', 'serial', (col) => col.primaryKey())
-    .addColumn('corporate_id', 'integer', (col) => col.notNull().references('corporates.id').onDelete('cascade'))
+    .addColumn('id', 'bigserial', (col) => col.primaryKey())
+    .addColumn('corporate_id', 'bigint', (col) => col.notNull().references('corporates.id').onDelete('cascade'))
     .addColumn('timestamp', 'timestamp', (col) => col.notNull())
     .addColumn('note', 'text')
     .addColumn('from_status', 'varchar(50)')
